+++ include/system/drivers/driver.hpp
@@ -12,6 +12,20 @@
 
 // typedef USBDevice USBDevice;
 
+class Driver
+    {
+    private:
+        char* Name;
+        char* Description;
+    public:
+        // static USBDevice* device; // Pointer to USBDevice
+        Driver(char* name = 0, char* description = 0);
+
+        char* GetDriverName();
+        char* GetDriverDescription();
+
+        virtual bool Initialize();
+    };
 
 
 static char* GetDriverName();
