--- include/system/drivers/usb/device.hpp
@@ -1,12 +1,0 @@
-#ifndef __ALINIX_KERNEL_DEVICE_HPP
-#define __ALINIX_KERNEL_DEVICE_HPP
-
-#include <alinix/types.h>
-#include <system/drivers/usb/usbdevice.hpp>
-#include <alinix/list.hpp>
-#include <system/drivers/usb/usbdriver.hpp>
-#include <system/drivers/usb/usbendpoint.hpp>
-
-class USBdevice;
-
-#endif
