--- kernel/c++/System/drivers/driver.cpp
@@ -1,7 +1,0 @@
-#include <system/drivers/driver.hpp>
-
-Driver::Driver(char* name, char* description)
-{
-    this->Name = name;
-    this->Description = description;
-}
